refactor: change entities to models
