+++ src/main/java/org/aster/models/Task.java
@@ -1,4 +1,4 @@
-package org.aster.entities;
+package org.aster.models;
 
 import java.time.LocalDateTime;
